Slayer area effects now allow for localization
